shorter posterior test
shorter time for the posterior testing
